requirements.txt
asgiref==3.5.0
black==21.12b0
click==8.0.3
colorama==0.4.4
Django==3.2.11
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==11.3.0
flake8==4.0.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.3
typing-extensions==4.0.1
